fix: Fixed indentation bug with comments
This fixes the bug where trying to indent when there's no comment
before the line we're trying to comment cause the line to indent
entirely to the left. This was caused by re-search-backward throwing
an error and ending the indentation calculation early.

This fix needs to be applied to the other two functions as well.